Add client/admin authentication and authorization

+++ backend/src/controllers/schedule.controller.ts
@@ -5,15 +5,28 @@ import { asyncMiddleware } from '../utils/async-middleware'
 export const scheduleController = Router()
 
 scheduleController.get('/', asyncMiddleware(async (req, res) => {
-  const schedules = await prisma.schedule.findMany({ include: { services: true } })
+  console.log({
+    where: { clientUuid: req.headers['x-client-uuid'] || '' },
+    include: { services: true }
+  });
+
+  const schedules = await prisma.schedule.findMany({
+    where: { clientUuid: req.headers['x-client-uuid'] || '' },
+    include: { services: true }
+  })
   res.json(schedules)
 }))
 
 scheduleController.post('/', asyncMiddleware(async (req, res) => {
   const { services = [], ...data } = req.body
+  const clientUuid = req.headers['x-client-uuid']
+
+  if (!clientUuid) return res.status(400).json({ error: `'x-client-uuid' header is not provided.`})
+
   const schedule = await prisma.schedule.create({
     data: {
       ...data,
+      clientUuid,
       services: { connect: services.map(id => ({ id })) }
     },
     include: { services: true }
@@ -23,6 +36,15 @@ scheduleController.post('/', asyncMiddleware(async (req, res) => {
 
 scheduleController.patch('/:id', asyncMiddleware(async (req, res) => {
   const { services, ...data } = req.body
+  const exists = await prisma.schedule.findFirst({
+    where: {
+      id: Number(req.params.id),
+      clientUuid: req.headers['x-client-uuid'] || ''
+    }
+  })
+
+  if (!exists) return res.status(404).json({ error: 'Not found' })
+
   const schedule = await prisma.schedule.update({
     where: { id: Number(req.params.id) },
     data: {
@@ -36,6 +58,15 @@ scheduleController.patch('/:id', asyncMiddleware(async (req, res) => {
 }))
 
 scheduleController.delete('/:id', asyncMiddleware(async (req, res) => {
+  const exists = await prisma.schedule.findFirst({
+    where: {
+      id: Number(req.params.id),
+      clientUuid: req.headers['x-client-uuid'] || ''
+    }
+  })
+
+  if (!exists) return res.status(404).json({ error: 'Not found' })
+
   await prisma.schedule.delete({ where: { id: Number(req.params.id) } })
   res.json({ id: Number(req.params.id) })
 }))
